+++ lib/src/deact/deact_instance.dart
@@ -1,10 +1,16 @@
 part of deact;
 
+typedef AfterRender = void Function(Deact);
+
 /// This interface provides an API to the Deact application.
 abstract class Deact {
   /// Returns the render time of the last update to the
   /// DOM in milliseconds.
   num get lastRenderTimeMs;
+
+  /// A function to be called after the node hierarchy was
+  /// rendered.
+  AfterRender afterRender;
 }
 
 class _DeactInstance implements Deact {
@@ -14,6 +20,8 @@ class _DeactInstance implements Deact {
   DeactNode rootNode;
   @override
   num lastRenderTimeMs;
+  @override
+  AfterRender afterRender;
 
   _DeactInstance(this.selector) : logger = Logger('deact.$selector');
 }